todo-app/pyproject.toml
[tool.poetry]
name = "devops-course-exercise"
version = "0.1.0"
description = ""
authors = ["Becky Carter <becky.carter@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
pytest = "^6.0.1"
selenium = "^3.141.0"
pylint = "^2.5.3"
python-dateutil = "^2.8.1"
python-dotenv = "^0.14.0"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
setup-script = "cp -n .env.template .env"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[virtualenvs]
create = true
in-project = true